pixel shader and animation

+++ GDENG03_DX/vertexBuffer.h
@@ -11,7 +11,7 @@ public:
 
 	bool release();
 
-	bool load(void* list_vertices, UINT size_vertex, UINT size_list, void* shader_byte_code, UINT size_byte_shader);
+	bool load(void* list_vertices, UINT size_vertex, UINT size_list, void* shader_byte_code, size_t size_byte_shader);
 	UINT getSizeVertexList();
 
 private:
